- get tx with big value to create tx

+++ rpcserver/rpcfunction.go
@@ -477,7 +477,7 @@ func (self RpcServer) handleCreateTrasaction(params interface{}, closeChan <-cha
 	}
 
 	// param #2: list receiver
-	totalAmmount := uint64(0)
+	totalAmmount := int64(0)
 	receiversParam := arrayParams[1].(map[string]interface{})
 	paymentInfos := make([]*client.PaymentInfo, 0)
 	for pubKeyStr, amount := range receiversParam {
@@ -489,7 +489,7 @@ func (self RpcServer) handleCreateTrasaction(params interface{}, closeChan <-cha
 			Amount:         uint64(amount.(float64)),
 			PaymentAddress: receiverPubKey.KeyPair.PublicKey,
 		}
-		totalAmmount += paymentInfo.Amount
+		totalAmmount += int64(paymentInfo.Amount)
 		paymentInfos = append(paymentInfos, paymentInfo)
 	}
 
@@ -500,7 +500,7 @@ func (self RpcServer) handleCreateTrasaction(params interface{}, closeChan <-cha
 		for _, desc := range temp.Descs {
 			for _, note := range desc.GetNote() {
 				amount := note.Value
-				totalAmmount -= amount
+				totalAmmount -= int64(amount)
 			}
 		}
 		txData := temp
